+++ openmdao/core/parallel_group.py
@@ -42,3 +42,26 @@ class ParallelGroup(Group):
                         tree[n] = data
 
         return tree
+
+    def _ordered_comp_name_iter(self):
+        """
+        Yield contained component pathnames in order of execution.
+
+        For components within ParallelGroups, true execution order is unknown so components
+        will be ordered by rank within a ParallelGroup.
+        """
+        if self.comm.size > 1:
+            names = []
+            for s in self._subsystems_myproc:
+                if isinstance(s, Group):
+                    names.extend(s._ordered_comp_name_iter())
+                else:
+                    names.append(s.pathname)
+            seen = set()
+            for ranknames in self.comm.allgather(names):
+                for name in ranknames:
+                    if name not in seen:
+                        yield name
+                        seen.add(name)
+        else:
+            yield from super()._ordered_comp_name_iter()
